Update Morse.cpp
Morse Code Converter
Morse code is a code where each letter of the English alphabet, each digit, and various
punctuation characters are represented by a series of dots and dashes. Table 10-8
shows part of the code.
Write a C++ program that asks the user to enter a string, and then converts that string to
Morse code. (Use hyphens for dashes and periods for dots.)